Package up the project
This packages up `randomfpl` so that it can now be installed with (for
example):

```bash
pip install .
```

or in editor mode with dev requirements:

```bash
pip install -e '.[dev]'
```

The script can then be run by simply calling:

```bash
randomfpl
```

The important file here is the `pyproject.toml`. I've also refactored
the `main` function so that it includes the call to `asyncio`.